feat: Update dependencies and add lazy loading for fonts and images
Refactor _layout.tsx to include lazy loading for fonts using the @expo-google-fonts/roboto package. This improves the page load performance by loading fonts only when needed. Additionally, update the Layout component to display a Loading component while the fonts are being loaded. This commit aims to enhance the user experience and optimize the application's performance.

Add lazy loading for images in the Main component to further improve page load performance. This commit introduces lazy loading for images using the react-native-gesture-handler package. By loading images only when they are in the viewport, the application can reduce the initial load time and improve the overall performance.

This commit also includes updates to dependencies, such as updating expo-font to version 12.0.6 and expo-app-loading to version 2.1.1, to ensure compatibility and take advantage of the latest features and bug fixes.