feat: Refactor Calendar component to accept onDayPress prop
The Calendar component has been updated to accept an onDayPress prop, which is a function that will be called when a day is pressed. This allows for more flexibility in handling the selected date and performing actions based on it. This change improves the reusability and customization of the Calendar component.